refactor tests for queues
